Use fluent WebhookEventCollectionBuilderCmsFile for file based webhook events (instead of only template/partial view)
