+++ src/Umbraco.Core/Webhooks/Events/File/ScriptSavedWebhookEvent.cs
@@ -1,0 +1,25 @@
+using Microsoft.Extensions.Options;
+using Umbraco.Cms.Core.Configuration.Models;
+using Umbraco.Cms.Core.Notifications;
+using Umbraco.Cms.Core.Services;
+using Umbraco.Cms.Core.Sync;
+
+namespace Umbraco.Cms.Core.Webhooks.Events.File;
+
+[WebhookEvent("Script Saved")]
+public class ScriptSavedWebhookEvent : WebhookEventBase<ScriptDeletedNotification>
+{
+    public ScriptSavedWebhookEvent(
+        IWebhookFiringService webhookFiringService,
+        IWebhookService webHookService,
+        IOptionsMonitor<WebhookSettings> webhookSettings,
+        IServerRoleAccessor serverRoleAccessor)
+        : base(webhookFiringService, webHookService, webhookSettings, serverRoleAccessor)
+    {
+    }
+
+    public override string Alias => "scriptSaved";
+
+    public override object? ConvertNotificationToRequestPayload(ScriptDeletedNotification notification)
+        => notification.DeletedEntities;
+}
