+++ src/Umbraco.Core/Webhooks/Events/File/TemplateDeletedWebhookEvent.cs
@@ -4,7 +4,7 @@ using Umbraco.Cms.Core.Notifications;
 using Umbraco.Cms.Core.Services;
 using Umbraco.Cms.Core.Sync;
 
-namespace Umbraco.Cms.Core.Webhooks.Events.Template;
+namespace Umbraco.Cms.Core.Webhooks.Events.File;
 
 [WebhookEvent("Template Deleted")]
 public class TemplateDeletedWebhookEvent : WebhookEventBase<TemplateDeletedNotification>
