--- src/Umbraco.Core/Webhooks/Events/Stylesheet/StylesheetSavedWebhookEvent.cs
@@ -1,25 +1,0 @@
-﻿using Microsoft.Extensions.Options;
-using Umbraco.Cms.Core.Configuration.Models;
-using Umbraco.Cms.Core.Notifications;
-using Umbraco.Cms.Core.Services;
-using Umbraco.Cms.Core.Sync;
-
-namespace Umbraco.Cms.Core.Webhooks.Events.Stylesheet;
-
-[WebhookEvent("Stylesheet Saved")]
-public class StylesheetSavedWebhookEvent : WebhookEventBase<StylesheetSavedNotification>
-{
-    public StylesheetSavedWebhookEvent(
-        IWebhookFiringService webhookFiringService,
-        IWebhookService webHookService,
-        IOptionsMonitor<WebhookSettings> webhookSettings,
-        IServerRoleAccessor serverRoleAccessor)
-        : base(webhookFiringService, webHookService, webhookSettings, serverRoleAccessor)
-    {
-    }
-
-    public override string Alias => "stylesheetSaved";
-
-    public override object? ConvertNotificationToRequestPayload(StylesheetSavedNotification notification)
-        => notification.SavedEntities;
-}
